relocate utils and make them reusable
