+++ src/utils/get-user.decorator.ts
@@ -1,6 +1,5 @@
 import { createParamDecorator, ExecutionContext } from "@nestjs/common";
 import { User } from "src/auth/user.schema";
-// import { User } from "./user.entity";
 
 export const GetUser = createParamDecorator((data, ctx: ExecutionContext): User => {
   const req = ctx.switchToHttp().getRequest();
